DevTools: Sources: Right clicking in a gutter always gets line number
If you right clicked very precicsely in the gutter, you could hit a div
that was not a line number. Then target.textContent would be " ", and
the line number would be NaN.

This patch uses a more robust method to find the line number div.

BUG=none

Review-Url: https://codereview.chromium.org/2545623002
Cr-Original-Commit-Position: refs/heads/master@{#436768}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 156eb1d0e4ab6d29b532a44738e9e7cd83618e8a