Fix crash when referencing a registered non-inherited custom property
This patch fixes a crash that occurs when a registered non-inherited
custom property is referenced on an element which doesn't explicitly
have any registered non-inherited custom properties set. The code was
missing a null-check for the non-inherited custom properties object.
Note that elsewhere in the file this is not needed as the object is
guaranteed to exist when we are resolving such properties.

BUG=671484

Review-Url: https://codereview.chromium.org/2552163002
Cr-Original-Commit-Position: refs/heads/master@{#436880}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1bf1f4ca99b8ea3183c9aa358338e63d5de07b93