[css-grid] Pass Grid as argument to more methods & remove m_gridIsDirty
This is the follow up of crrev.com/2536703002, three more methods get the
Grid passed as argument. The remaining methods will access the Grid through
GridSizingData, but that will be another patch.

Apart from that, m_gridIsDirty was removed because it was always too
confusing. It was replaced by Grid's m_needsItemsPlacement which is much
more concise. The dirtyGrid() call was indeed only forcing a replacement of
the grid items.

BUG=627812

Review-Url: https://codereview.chromium.org/2541003003
Cr-Original-Commit-Position: refs/heads/master@{#436910}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8f9db3a8e2aff0307f677c3b29efc0162a80b5f