Embedding-CSP: Ports subsumption
When we check for subsumption, we need to make sure that even if ports
are different but are defaults for the specified protocols, we return
correct result. Consider:

A = http://example.com:80
B = https://example.com:443

Then A subsumes B.
Moreover, A is similar to B and their intersection should be
`https://example.com:443`.

BUG=647588

Review-Url: https://codereview.chromium.org/2556713002
Cr-Original-Commit-Position: refs/heads/master@{#436916}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9881be66406b729bfa518c4aaf54430ed5bc0683