Media Controls: don't try to cut the controls when the element is fullscreen.
During the fullscreen animation, the reported position of the media
element will be greater than 0 and it is sometimes not reset to 0 as it
should be, leaving the controls cut on the right side.

This a work around a fullscreen issue in order to avoid a regression.

BUG=663680
R=foolip@chromium.org

Review-Url: https://codereview.chromium.org/2550423003
Cr-Original-Commit-Position: refs/heads/master@{#436937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4110ecc69aca6f4fb7b12ed9118844e8d883014