Speculative fix for crash in canvas hibernation
Based on crash report analysis it seems that when a canvas is
attempting to enter hibernation, sometimes the 2d rendering context
has a null pointer in its state stack.  Upon inspection of code
BaseRenderingContext2D, it seems impossible that a null pointer
would ever be inserted into m_stateStack. The hypothesis for this
speculative fix is a race condition on object destruction.
We speculate that the hibernation task is being invoked after the
canvas element has been garbage collected, but the ImageBuffer
has not, which would cause ImageBuffer::m_client to be used after
being freed. To fix this we reset the ImageBuffer's client pointer
in the HTMLCanvasElement's finalizer.

BUG=648510

Review-Url: https://codereview.chromium.org/2549353004
Cr-Original-Commit-Position: refs/heads/master@{#436969}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9135fa05fee25fe398b1b495f8b8359cad55db95