Prioritize hyphenations over break-word/break-all
When these mid-word break properties are used together, this patch
tries to hyphenate first.

This behavior matches to WebKit, and to Gecko/Edge for break-word. The
priority isn't defined in the spec, currently an open issue[1]. But for
break-word, 4 browsers are interoperable with this change.

Also fixes one of failing cases in fast/text/basic/015.html that is
related with additionalWidthFromAncestors when break-word.

[1] https://github.com/w3c/csswg-drafts/issues/791

BUG=671341

Review-Url: https://codereview.chromium.org/2560453002
Cr-Original-Commit-Position: refs/heads/master@{#436979}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0922f146b8701a3609d57e0fb11b3eb9d35ceb47