Adding an experimental flag to block autoplay with sound in cross-origin iframes
This CL adds an experimental flag to block autoplay with sound in
cross-origin iframes on Desktop. The experimental flag is
disabled by default, and will override
the "MediaPlaybaRequiresUserGesture" flag.

Currently, the default behavior is unchanged, and a follow-up CL
will collect metrics pretending the flag is on.

Design doc:
https://docs.google.com/a/google.com/document/d/1g05pahL0py0TSWIDwli1MLJiRuuJoleitl572DGGESs/edit?usp=sharing

BUG=665456

Review-Url: https://codereview.chromium.org/2496593002
Cr-Original-Commit-Position: refs/heads/master@{#437254}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df72ed087b3083bdea9b882bf473f8a9f4dc1276