Track intersection of HTMLMediaElement with root frame.
This is the intersection of interest; previously, if a media element
was in an iframe, it would track its intersection with the iframe
boundary.

Note that the localFrameRoot() code is temporary; a follow-up patch
which refactors IntersectionGeometry will change this argument to
nullptr.  I'm putting this one-line change in separate patch to isolate
the behavioral change.

R=xjz@chromium.org
BUG=643964

Review-Url: https://codereview.chromium.org/2558633003
Cr-Original-Commit-Position: refs/heads/master@{#437319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f57a6a8e8b77b55410f1c16340fcc1608298e4f4