Disable form elements in MHTML
MHTML is loaded in full sandboxing mode and thus the form cannot
be submitted and the script execution is disabled. So we want to
disable all form elements in order to give a clear indication to
the user that the form cannot be worked on.

BUG=669325
TEST=new tests added

Review-Url: https://codereview.chromium.org/2560693002
Cr-Original-Commit-Position: refs/heads/master@{#437362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea856e3748ce4836299faf081f96c0b37ee923b6