Rename HTMLScriptRunner to HTMLParserScriptRunner and add comments.
This class is only responsible for scripts inserted and executed due to
HTMLDocumentParser, so it's renamed to reflect this (and hopefully make its
relationship to ScriptRunner and ScriptLoader clearer).

execute is renamed to processScriptElement, because it contains the logic
for processing a script end tag (from the spec), which does not necessarily
execute the script synchronously.

isPendingScriptReady is renamed to isParserBlockingScriptReady, since that
is the particular pending script it means, as opposed to
m_scriptsToExecuteAfterParsing.

There is still some more minor inconsistency in the use of the term
"parser-blocking script" as opposed to the spec's "parsing-blocking script",
but this would require changes to more connected classes to fix.

executeParsingBlockingScript is inlined into executeParsingBlockingScripts,
since both methods are small but have confusingly similar names.

ASSERT is changed to DCHECK because check-webkit-style insists.

This should cause no functional changes.

Review-Url: https://codereview.chromium.org/2559803002
Cr-Original-Commit-Position: refs/heads/master@{#437363}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 799df4e1bd56ebe1b03e0bfdc01d244e3e54db47