Devirtualize Frame::domWindow().
Originally, this was a virtual, but it meant that LocalFrame::domWindow
could not return a LocalDOMWindow, due to the circular dependencies
between the LocalFrame and LocalDOMWindow header. As a workaround,
LocalFrame exposed a special localDOMWindow getter. However, this leads
to code that reads very LocalFrame-centric. Since the ultimate goal of
OOPIF integration is to integrate seamlessly, it would be nice to
remove the "local" prefix from the getter.

Since this function doesn't need to be virtual, just de-virtualize it
and have the getters of derived classes return the appropriate type.

BUG=none

Review-Url: https://codereview.chromium.org/2562323002
Cr-Original-Commit-Position: refs/heads/master@{#437887}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fd3cea4f35b5b831cef689f0720f13d8f22ba103