DCHECK in BudgetService getCost and reserve if an invalid type is supplied.
If an incorrect type is provided to the call, the generated IDL code will
reject with a TypeError. This patch removes code in the service
implementation which was checking whether the type was invalid and returning
a NotSupportedError. The only way this code could be hit would be if there
was a programming error, so it should be a DCHECK.

BUG=672113

Review-Url: https://codereview.chromium.org/2570463003
Cr-Original-Commit-Position: refs/heads/master@{#437897}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c774d5e31f3d0333dfdf921c0befcc89d1bd1259