Fix gmock leak issue in AutoplayUmaHelperTest
AutoplayUmaHelperTest currently uses GMock for mock
AutoplayUmaHelper in order to check if events are correctly
received. However this will make the test flaky since
MockAutoplyUmaHelper is garbage-collected but GMock expect it to
be destroyed when the test finishes.

This CL marks MockAutoplayUmaHelper as AllowLeak which fixes the
issue.

BUG=672831

Review-Url: https://codereview.chromium.org/2562683004
Cr-Original-Commit-Position: refs/heads/master@{#437939}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8906ce8b964bf3431b891bc9eaaca54ad8981f7d