Add support for caching and invalidating SVG marker transforms
This patch adds a cache of the SVG marker transform, similar to the pattern
used in LayoutSVGViewportContainer. This should be a small perf win because
we no longer re-calculate the transform when painting each marker instance.
With this new cache logic, invalidation logic has been added which includes
a fix for paint property under-invalidation (marker-viewBox-changes.svg).

BUG=645667,600618

Review-Url: https://codereview.chromium.org/2565943002
Cr-Original-Commit-Position: refs/heads/master@{#437993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d798a7a82ead340d6206380d64e310c222416e62