Rename blink::Document::ReadyState enum to DocumentReadyState.
The rename is needed to avoid a naming collision after changing from
Blink to Chromium naming style.  Right now we have a |ReadyState| type
and a |readyState| accessor method (differing by case of the first
character); after a naive rename by the rewrite_to_chrome_style tool we
would end up with |ReadyState| being the name of both the type and the
accessor method (with both living in the same namespace).

Prepending a "get" prefix to the name of the accessor method is the
workaround that fits into the guidance on the recommended
post-Blink-to-Chromium-rename style suggested by esprehn@ in
https://crbug.com/582312#c17:
- Getters favor not using "Get", ex. FirstChild()
- Unless the type name conflicts, in which case you can either rename
  the type if it's easy and makes sense, or add "Get", ex. GetContext().

FWIW, this rename brings the name of the enum closer to the name of the
enum in the WHATWG spec.

BUG=582312

Review-Url: https://codereview.chromium.org/2567943004
Cr-Original-Commit-Position: refs/heads/master@{#438018}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b090f5776dec4e955552c8e46f2bed21e3a9cff