Changed EOverflowAnchor to an enum class and renamed its members
Changed EOverflowAnchor to an enum class and gave it an unsigned
underlying type.

Changing it to an enum class enforces better namespacing and code
practices. Adding the unsigned underlying type is pre-work for when the
class is eventually stored as an enum bitfield (it would be done in this
patch, except a presubmit warning already exists that prevents that. The
presubmit warning needs to be updated before that change can occur.)

This is also pre-work to move EOverflowAnchor to be generated in
ComputedStyleBase.

BUG=628043
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2561973002
Cr-Original-Commit-Position: refs/heads/master@{#438072}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad2fe04892fed522eb67cd054f8696a5cfb1ba3f