Remove attribute filters from .tmpl files.
We rarely handle all attributes in a loop.
Instead, we have filters to iterate only on accessors,
on data type attributes, so on.
It is very confusing to have all attributes in a
variable in .py and filter them in .tmpl files.

This CL applies such filters in pre-process in .py
and uses more descriptive names in template contexts.
It keeps 'attributes' for some usages, which are
to be replaced.

This is a refactoring in template engine, so it
does not change generated code.

BUG=650150

Review-Url: https://codereview.chromium.org/2568403002
Cr-Original-Commit-Position: refs/heads/master@{#438099}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b9858514e087acbcf12846f3cd77220ba964f14