Updated the return of GetBudget to a sequence
The proposed spec for BudgetAPI defines the return value for getBudget
to be a Promise that resolves with a sequence<BudgetState> while the
current interface used a FrozenArray.

This patch updates the code to match the proposed spec. The implementation
works with either version, and this is covered by existing LayoutTests.

BUG=672112

Review-Url: https://codereview.chromium.org/2571573002
Cr-Original-Commit-Position: refs/heads/master@{#438127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 44433c83d095d50a6c46fa6a912156bfbfada092